selinux-testsuite: add labeled ipsec tests
Add tests for labeled IPSEC.  These are identical to the tests
for NetLabel, i.e. they exercise labeled networking over loopback,
including testing of SO_PEERSEC, SCM_SECURITY, and peer recv permission.

One issue with these tests is that even after flushing the ipsec
SAD and SPD entries, the kernel continues to perform the peer checks
which can cause denials for e.g. NFS once the test policy is unloaded.
This suggests some problem with the selinux xfrm ref counting or perhaps
delayed freeing of the entries.  To compensate, we allow peer permissions
for the existing connect tests so that they do not fail if we run the
testsuite a second time without an intervening reboot.

Another issue is that unlike the NetLabel case, we do not get an ICMP
error back upon the peer recv denial so the client has to wait for
a connection timeout, although it does ultimately fail and therefore
the test does pass.  To avoid hanging for a while on this test, we
explicitly set the timeout value for the connect() call.

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>