selinux-testsuite: move unix_socket and mmmap tests out of SUBDIRS_COMMON
RHEL 4 and 5 have forked test policies under policy/redhat, so
we cannot add new tests to SUBDIRS_COMMON unless we want to port the
policy rules for those tests to the rhel 4 and 5 test policies
as well.  For now, just move the unix_socket and mmap tests out of
SUBDIRS_COMMON to SUBDIRS so that they are only run on Fedora and
RHEL 7.  We might also be able to add them to the RHEL6 SUBDIRS
definition as it shares the same test policy but that has not been
tested.

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>