selinux-testsuite: enable running new tests on RHEL6/7
RHEL 6 and 7 do not include ipsec-tools, so rewrite the ipsec
tests to use ip xfrm for loading the test configuration rather than
setkey.

RHEL 6 unlabelednet policy module only includes rules for the domains
in the RHEL 6 policy, not a rule written on the domain attribute, so
we have to explicitly allow unlabeled send/recv for the test
domains in order for the tests to succeed there.

RHEL 6 uses its own SUBDIRS definition to exclude the nnp test, so
add the mmap, unix_socket, and inet_socket tests to its list.

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>