selinux-testsuite: ipsec-load: Derive state context from current context.
Change ipsec-load to derive the state context from the current context
rather than assuming that the caller always has a given user/role/range.
This is similar to logic elsewhere throughout the tests that only
sets the type to one of the test domains/types and otherwise inherits
the user/role/range from the caller (for a process) or a related object
(for a file).  This keeps the test independent of whether it is called
from e.g. an unconfined user shell vs a sysadm user shell or from
targeted vs mls policy.

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>