Build the SELinux testsuite on systems using the Reference Policy
Sorry about that ! I did not test it properly because I was hitting a
problem with the policy...

The following one should work, although it's slightly noisy...

Fix the Makefile in the policy subdirectory of the SELinux testsuite
so that it builds correctly on the standard Reference Policy which
does not have the same booleans available on Red Hat.

Previous versions were not working within the Makefile.

Signed-off-by: Guido Trentalancia <guido@trentalancia.net>