selinux-testsuite: Add tests for extended socket classes.
Add tests for the extended_socket_class policy capability.
This change includes the following tests:
- Test that ICMP datagram sockets are mapped to the new icmp_socket
class and not to rawip_socket for both IPv4 and IPv6.

- Test that SCTP stream and seqpacket sockets are mapped to the
new sctp_socket class and not to rawip_socket for both IPv4 and IPv6.

- Test that Bluetooth sockets are mapped to the new bluetooth_socket
class and not to socket.

- Test that AF_ALG sockets are mapped to the new alg_socket class
and not to socket.

The tests are only run if the extended_socket_class policy capability
is present and enabled in the kernel and the base policy, and only if
the new classes are defined in the base policy.  This avoids breaking
the testsuite on systems with older kernels, older policies, or
policies that do not enable the policy capability.

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>