selinux-testsuite: fix ptrace test script
The ptrace test script was doing a chmod +x wait, which wouldn't work
when run from the top-level directory.  Use $basedir to fully qualify
the pathname.  This just happened to "work" when there was a wait
test, but it was then being applied to the wait subdirectory, not the
wait script.

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>