selinux-testsuite: fix nnp test for RHEL7
RHEL7.3 backported the support for bounded transitions under
NO_NEW_PRIVS.  Since the kernel version comparison is against
the upstream kernel version that first included this support,
the test was failing on RHEL7.  Adjust the logic so that we do
not treat it as an error if earlier kernels correctly return
success (i.e. they contain the back-ported patch).

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>