selinux-testsuite: allow more instances of map permission
The upstream patch defining map permission for refpolicy
("refpolicy: Define and allow map permission") only adds
map permission to mmap_file_perms and exec_file_perms and to
specific interfaces where it is required.  Consequently, we need
to allow more instances of map permission in the file and mmap
test policies to avoid incorrect denials of map permission
(either ones that cause a test that should succeed to fail,
or ones that would prevent the test from reaching a later
permission check that is being tested).

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>