tools: add a syntax checker/fixer
Add a simple script that uses astyle to check, and optionally fix,
the syntax of the test sources.  At the moment it is limited to C
but it can easily be augmented to support Perl in the future.

We also add a new make target, 'check-syntax', to make it easier for
developers to check their changes.

Signed-off-by: Paul Moore <paul@paul-moore.com>