selinux-testsuite: Add tests for transitions under NNP/nosuid
Duplicate the existing tests for transitions under NNP for
transitions on a nosuid mount, and then augment both the NNP
and nosuid tests to also test the new support for allowing
transitions based on nnp_transition and/or nosuid_transition
permission if the nnp_nosuid_transition policy capability is
enabled.  Test NNP and nosuid independently and together.

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>