selinux-testsuite: only add the infiniband tests to SUBDIRS if they are enabled
It was difficult to tell if the InfiniBand tests were actually being
run as they were always added to SUBDIRS regardless of the
configuration; this patch should make this more obvious.

It is worth nothing that most of the changes to the tests themselves
are simply due to reformatting (./tools/check-syntax -f) due to the
removal of the enabled check.

Signed-off-by: Paul Moore <paul@paul-moore.com>