overlay: Fix bad entry point test for context mounts
Current bad entry point test for context mounts does not make much sense.
During the test we are mounting overlay with context=...rwx_t. And that
means process will see this label on overlay inode and that should allow
entry.

We are expecting entry to fail. But, given process is seeing rwx_t, and
as per policy entrypoint into that is allowed. So this test case in
current form does not make much sense for context mounts.

Why it works currently, because selinux is actually checking the label
of lower file (ro_t) instead of label of overlay inode (rwx_t) and that's
why entrypoint fails. But this is wrong expectations.

So remount overlay with context=...ro_t and run the test and now it
should fail. (lable ro_t does not have entrypoint permission).

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>