selinux-testsuite: fix client/server race conditions
There is a race condition in the SCTP tests that causes failures when
the SCTP server program takes too long to get past the listen() call. In
such case the client attempts to connect to the port when noone is
listening yet, leading to a test failure. Similat race conditions are
possible in the unix_socket, inet_socket, and binder tests, although
these do not seem to trigger in practice so easily.

This situation can be easily reproduced by adding sleep(1) before the
listen() call in tests/sctp/sctp_server.c.

Fix this by having the server programs write something to a fifo after
successfully calling listen() to inform the test script that it can run
the client.

Another possible race condition is due to the 'kill TERM, ...' calls,
which may not terminate the server quickly enough, leading to the new
server failing to bind to the port (which is still being used by the old
server). Fix this by sending the KILL signal rather than TERM (there is
really no point in being graceful here).

Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>