selinux-testsuite: Update binder test applications
Replace binder_test.c with separate manager, client and service provider.
This works in the same way as a service provider/client interacts
with a service manager in the Android world. It passes the service
providers binder file descriptor to the client for the impersonate
permission check.

Also added tests for Dynamically Allocated Binder Devices and passing
the sender SELinux security context on binder transactions.

Note that the tests require a minimum kernel of 4.16, else some tests may
fail. To run successfully the "binder: Add thread->process_todo flag"
patch may be required that is available from:
https://lore.kernel.org/patchwork/patch/851324/
This patch has been backported to some earlier kernels.

Signed-off-by: Richard Haines <richard_c_haines@btinternet.com>
Signed-off-by: Paul Moore <paul@paul-moore.com>