tests: wait for servers to finish before continuing
When we are starting a new server after killing an old one, there is a
possible race condition if the old server takes too long to tear down
and the new server fails to bind.

I've seen this happen in practice with the sctp test, but let's fix this
also in other tests where it makes sense logically.

Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>