README: update kernel-devel package in dnf command
The README.md file in the selinux-testsuite states that kernel-devel
package is needed. However, for everything to work you need the
kernel-devel package specific to your kernel version.
Update the dnf command to include uname -r, so that a copy + paste
of the dnf command, results in a system with dependencies filled
for recent Fedora distro's like Fedora 32.

Signed-off-by: William Roberts <william.c.roberts@intel.com>
Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>