Migrate CI to GH Actions
As discovered by Nicholas Iooss [1], GH Actions now supports
HW-accelerated Linux VMs at least on macOS runners [2], so we can
finally migrate the CI away from Travis.

The new check uses Vagrant to provision the Fedora VM, as this is the
simplest way to get it to work on macOS. Unfortunately there is no
Rawhide image provided in the official Vagrant repos, so for now the
testsuite is run only on Fedora 33. I'd like to add Rawhide in the
future, but let's start with the low-hanging fruit.

[1] https://github.com/SELinuxProject/selinux/commit/316a4f89dd502d6f2b072f76af16de320c1cf673
[2] https://github.com/actions/virtual-environments/issues/433#issuecomment-618872784

Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>