travis: update the kernel and reboot before running the suite
While the stable release cloud image always ships with a kernel which
has the matching -modules and -devel packages available in the "fedora"
repo, on development releases the nightly images may be behind the
latest repo content and thus may fail to install kernel-modules and
kernel-devel.

Fix it by installing the latest available the kernel and booting it
before running the testsuite. This unfortunately adds around 3-4 minutes
to the Travis build time, but there seems to be no better way to make
the Rawhide test work.

Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>