travis: skip kernel update+reboot if it's not necessary
Only do it when we can't install the kernel-modules package for the
running kernel to make the CI run faster. Now that rawhide images are
updating regularly again, it is rarely necessary.

Also don't run `dnf clean all -y` before installing dependencies. There
really is no point and it just slows things down when we do need to
update the kernel.

Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>
Acked-by: William Roberts <bill.c.roberts@gmail.com>