travis: handle the case of multiple rawhide images
For whatever reason, some mirrors contain more than one image version in
the directory, causing the test to bail out. This happens only sometimes
and restarting the Travis job usually helps, but it's better to adapt
the script to handle this case.

Tweak the script to pick the lexicographically last image and exit with
error only if there is no image at all.

Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>