tests/userfaultfd: handle __NR_userfaultfd not being defined
On some old kernels (think RHEL-7) __NR_userfaultfd may not be defined
on certain arches, even though the <linux/userfaultfd.h> header is
available. To avoid build errors in such environments, abstract the
userfaultfd syscall into a helper function and make it fail with ENOSYS
when __NR_userfaultfd is not defined.

Reviewed-by: Lokesh Gidra <lokeshgidra@google.com>
Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>