Add AF_VSOCK SOCK_STREAM client/server tests
If AF_VSOCK and vsock_loopback are supported by the system, run tests
that exchange a byte of data between a client and a server listening on
VMADDR_CID_LOCAL and a random port.

Various permissions are removed from the client/server between runs and
it is checked that the corresponding syscalls returned error.

A newly created vsock_socket inherits the SID of the current process and
it is tested that the vsock_socket returned by accept() inherits the same
SID from its parent.

SOCK_DGRAM is not tested as it is only supported in the VMCI transport.

Signed-off-by: David Brazdil <dbrazdil@google.com>
Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>