ci: test also on F34 images
Change the CI configuration to test also on an F34 image in addition to
F33 and also test kernel-secnext on F34 instead of F33.

We could of course test just on F34, but having an older release in the
test matrix should help ensure that the testsuite will still work on
older kernel/userspace versions as well, so I'd rather keep both.

Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>