tests: exclude vsock_socket test where it wouldn't build
The test requires the <linux/vm_sockets.h> header to be available and
contain th definition for VMADDR_CID_LOCAL, so wrap it in a check for
these preconditions.

Likewise, the vsock_socket test policy requires the vsock_socket class
to be defined in the base policy, so wrap the policy bit and test in a
check for this, too.

Without this patch the testsuite wouldn't build on older distros such as
RHEL-7.

Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>