tests/perf_event: don't assume CPU#0 is online
The perf_event test currently uses hard-coded 0 as the CPU index.
Howver, if CPU#0 happens to be offline, the test then fails because
perf_event_open(2) returns -ENODEV.

To fix this, try to find the index of the first online CPU and use that
instead of the hard-coded value.

Verified to work well on a machine with the first 8 CPU cores offline.

Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>