tests/sctp: add client peeloff tests
Currently the testsuite only verifies that SCTP peeloff works on the
server side. However, it can just as well be used on the client side as
well, which isn't being tested (and actually is buggy at the time of
writing).

To correct this, add a simple SCTP peeloff client and a couple tests
that verify that the client-side peeloff works well with SELinux.

The new tests are enabled on kernels 5.18+, where the fixes were
introduced. While the fixes are now queued also for some stable streams,
it's much simpler to just check the version against 5.18 and skip the
test for all older kernels.

Tested-by: Richard Haines <richard_c_haines@btinternet.com>
Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>