tests/module_load: detect the linker to use for module build
Similar to the compiler, matching the linker used to compile the kernel
is also important for an external kernel module build. Add code to
detect the linker from the kernel config similar to the existing
compiler detection.

Speicifically, this fixes kernel module builds under kernels built with
clang and with LTO enabled.

Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>