tests/sctp: remove assumptions in the SCTP tests
Rework the SCTP tests slightly to remove two assumptions which are
not always guaranteed to be true (below).  This should have not any
affect on the current test suite or released kernels, but it will
help ensure that the test suite continues to work with upcoming
kernel releases.

 * Do not rely on IP options attached to a socket.  Depending on the
   kernel configuration, the on-the-wire packet labels may be
   generated on a per-packet basis as opposed to a per-socket basis.

 * Ensure the kernel's SCTP association state is properly reset by
   restarting the SCTP test server between tests which require it.

[-- further explanation from OM below --]

Before this patch, the test:
  * On the client side: extracted the outbound label from the
    socket-level IP options and expected to receive this from the
    server.
  * On the server side: extracted the outbound label from the
    socket-level IP options and sent it to the client.

This patch corrects the test to compare the peer context of the client
as it appears to the server to the expected value (in some cases derived
from the client's process context), which makes much more sense.

The restarting of the seq servers is needed because the server now
relies on the socket-level peer context, which in the case of
SOCK_SEQPACKET gets initialized to the first peer's context and then
stays the same for the rest of the socket's lifetime.

Signed-off-by: Paul Moore <paul@paul-moore.com>
[OM: added more details into the commit description]
[OM: fixed tools/check-syntax complaints]
Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>